backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
pg8000==1.29.8
psycopg2-binary==2.9.10
redis==4.6.0
PyJWT==2.8.0
cryptography==41.0.4
ecdsa==0.18.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0
email-validator==2.0.0
SQLAlchemy==2.0.21
requests==2.31.0
flask-restx==1.3.0
psutil==5.9.5
flask-compress==1.14